ecs_test的多重高斯模糊和全局泛光

+++ tests/ecs_test/glsl/post_vs.glsl
@@ -1,5 +1,5 @@
-#version 430 core
-precision lowp float;
+#version 320 es
+precision highp float;
 
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec2 aTexCoords;
